Remove previous code zero-ing out NAs
Jeanette explained what was really needed here which is to fill in a mix of zeros and NAs in the filled-in (added) columns depending on what values make sense. Zeroes go where the fish would have returned by that row's brood year and NAs go where the fish would not have returned yet.